If statements replaced
Replaced the if statements with a switch case statement. The switch case statement is still large but uses fewer lines than the if statement to get the same result. I wanted to make this change because I looked at the if statements and noticed how many there were and I remembered switch case. This took some time to do but I think it looks better now.